+++ jerboa/src/main/java/com/boundless/jerboa/animation/Coordinate.kt
@@ -4,3 +4,25 @@ data class Coordinate(
         val x: Double,
         val y: Double
 )
+
+fun Coordinate.isInBounds(bounds: Rectangle): Boolean {
+    if (x > bounds.left &&
+        x < bounds.right &&
+        y > bounds.top &&
+        y < bounds.bottom
+    ) {
+        return true
+    }
+
+    return false
+}
+
+fun Coordinate.move(
+        angleOfMovementInRadians: Double,
+        distance: Double
+): Coordinate {
+    return Coordinate(
+            x = x + (distance * Math.cos(angleOfMovementInRadians)),
+            y = y + (distance * Math.sin(angleOfMovementInRadians))
+    )
+}