+++ jerboa/src/main/java/com/boundless/jerboa/animation/Direction.kt
@@ -1,0 +1,8 @@
+package com.boundless.jerboa.animation
+
+enum class Direction {
+    UP,
+    LEFT,
+    DOWN,
+    RIGHT
+}