+++ jerboa/src/main/java/com/boundless/jerboa/animation/Rectangle.kt
@@ -26,3 +26,43 @@ data class Rectangle(
         return false
     }
 }
+
+fun Rectangle.translateDown(translation: Float): Rectangle {
+    val newCenterCoordinate = this.center.copy(y = this.center.y + translation)
+    return this.copy(center = newCenterCoordinate)
+}
+
+fun Rectangle.translateUp(translation: Float): Rectangle {
+    val newCenterCoordinate = this.center.copy(y = this.center.y - translation)
+    return this.copy(center = newCenterCoordinate)
+}
+
+fun Rectangle.translateLeft(translation: Float): Rectangle {
+    val newCenterCoordinate = this.center.copy(x = this.center.x - translation)
+    return this.copy(center = newCenterCoordinate)
+}
+
+fun Rectangle.translateRight(translation: Float): Rectangle {
+    val newCenterCoordinate = this.center.copy(x = this.center.x + translation)
+    return this.copy(center = newCenterCoordinate)
+}
+
+fun Rectangle.isOffBottomOfBounds(bounds: Rectangle): Boolean {
+    if (this.top > bounds.bottom) {
+        return true
+    }
+
+    return false
+}
+
+fun Rectangle.isInBoundsOf(rectangle: Rectangle): Boolean {
+    if (this.left > rectangle.left &&
+        this.right < rectangle.right &&
+        this.top > rectangle.top &&
+        this.bottom < rectangle.bottom
+    ) {
+        return true
+    }
+
+    return false
+}