+++ jerboa/src/main/java/com/boundless/jerboa/platform/BasicMusicPlayer.kt
@@ -1,0 +1,10 @@
+package com.boundless.jerboa.platform
+
+interface BasicMusicPlayer {
+    fun load(
+            resource: Int,
+            looping: Boolean = false
+    )
+    fun start()
+    fun stop()
+}