+++ jerboa/src/main/java/com/boundless/jerboa/ui/SectionViewState.kt
@@ -1,0 +1,9 @@
+package com.boundless.jerboa.ui
+
+/**
+ * @param sectionId used to distinguish which section has been clicked
+ * when there are multiple of the same section
+ */
+interface SectionViewState {
+    val sectionId: String
+}